requirements.txt
# Core scientific packages
numpy>=1.20.0
pandas>=1.2.0
scikit-learn>=0.24.0
joblib>=1.0.0

# Web requests and HTML parsing
requests>=2.25.0
beautifulsoup4>=4.9.0

# Command line and environment
argparse>=1.4.0
python-dotenv>=0.17.0

# Email handling
email>=0.9.0

# Misc
pathlib>=1.0.1
re>=2.2.1